Refactor: 타입 일치하게 수정

+++ backend/src/main/java/com/team6/issue_tracker/domain/page/dto/IssuePageResponse.java
@@ -10,8 +10,9 @@ import java.util.List;
 
 @Data @Builder
 public class IssuePageResponse {
-    private Integer openIssueCount;
-    private Integer closedIssueCount;
+
+    private Long openIssueCount;
+    private Long closedIssueCount;
     private Integer page;
     private Integer openIssueMaxPage;
     private Integer closeIssueMaxPage;
